[REFACTOR] Extract IndonesiaMap component from page.tsx into a separate file
